#96 - separando arquitetura de arquivos
