+++ backend/main.py
@@ -1,57 +1,18 @@
-from fastapi import FastAPI, APIRouter, WebSocket, WebSocketDisconnect
+from fastapi import FastAPI, APIRouter
 from fastapi.middleware.cors import CORSMiddleware
 from api.routers import routes
 from dotenv import load_dotenv
 import uvicorn
 import os
-from fastapi.responses import HTMLResponse
 
 from api.websocket.connection_manager import ws_manager
-from api.schemas.websockets import WSMessage, WSMessageAction
 
-from api.websocket.connection_manager import ws_manager as manager
+from tests.websocket_test_chat import router as websocket_test_routes
+from api.websocket.websocket_endpoint import router as websocket_routes
 
 load_dotenv()
 
 app = FastAPI()
-
-html = """
-<!DOCTYPE html>
-<html>
-    <head>
-        <title>Chat</title>
-    </head>
-    <body>
-        <h1>WebSocket Chat</h1>
-        <h2>Your ID: <span id="ws-id"></span></h2>
-        <form action="" onsubmit="sendMessage(event)">
-            <input type="text" id="messageText" autocomplete="off"/>
-            <button>Send</button>
-        </form>
-        <ul id='messages'>
-        </ul>
-        <script>
-            var client_id = Date.now()
-            document.querySelector("#ws-id").textContent = client_id;
-            var ws = new WebSocket(`ws://localhost:8000/ws/${client_id}`);
-            ws.onmessage = function(event) {
-                var messages = document.getElementById('messages')
-                var message = document.createElement('li')
-                var content = document.createTextNode(event.data)
-                message.appendChild(content)
-                messages.appendChild(message)
-            };
-            function sendMessage(event) {
-                var input = document.getElementById("messageText")
-                ws.send(input.value)
-                input.value = ''
-                event.preventDefault()
-            }
-        </script>
-    </body>
-</html>
-"""
-
 
 app.add_middleware(
     CORSMiddleware,
@@ -65,53 +26,13 @@ app.add_middleware(
 for route in routes.routes:
     app.include_router(route)
 
+app.include_router(websocket_test_routes)
+app.include_router(websocket_routes)
+
 
 @app.get("/")
 async def index():
-    # message = WSMessage[WSMessageAction.DeletedComment](
-    #     action=WSMessageAction.UpdatedTweetLike,
-    #     body=schemas.WSTweetLikeUpdated(
-    #         isLiked=True, tweetLike=schemas.TweetLikeResponseBody(tweetId=1, userId=2, username="teste")
-    #     ),
-    # )
-    # print(message)
-    # await ws_manager.broadcast(message, current_user.id)
-    return HTMLResponse(html)
-
-
-# class ConnectionManager:
-#     def __init__(self):
-#         self.active_connections: list[WebSocket] = []
-
-#     async def connect(self, websocket: WebSocket):
-#         await websocket.accept()
-#         self.active_connections.append(websocket)
-
-#     def disconnect(self, websocket: WebSocket):
-#         self.active_connections.remove(websocket)
-
-#     async def send_personal_message(self, message: str, websocket: WebSocket):
-#         await websocket.send_text(message)
-
-#     async def broadcast(self, message: str):
-#         for connection in self.active_connections:
-#             await connection.send_text(message)
-
-
-# manager = ConnectionManager()
-
-
-@app.websocket("/ws/{client_id}")
-async def websocket_endpoint(websocket: WebSocket, client_id: int):
-    await manager.connect(websocket, client_id)
-    try:
-        while True:
-            data = await websocket.receive_text()
-            await manager.send_personal_message(f"You wrote: {data}", websocket)
-            await manager.broadcast(f"Client #{client_id} says: {data}", client_id)
-    except WebSocketDisconnect:
-        await manager.disconnect(websocket)
-        await manager.broadcast(f"Client #{client_id} left the chat", client_id)
+    return {"api_status": "ok"}
 
 
 if __name__ == "__main__":
